# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SiapPA API",
        "description": "API documentation for SiapPA application",
        "contact": {
            "email": "support@siappa.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/artikels": {
            "get": {
                "tags": [
                    "Artikels"
                ],
                "summary": "Get list of artikels",
                "description": "Returns a list of artikels ordered by the latest first",
                "operationId": "getArtikelsList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Artikels"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Failed to fetch data"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error message details"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Artikels"
                ],
                "summary": "Create a new artikel",
                "description": "Creates a new artikel and returns the created artikel data",
                "operationId": "createArtikel",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "judul": {
                                        "description": "Judul artikel, maksimum 255 karakter",
                                        "type": "string",
                                        "example": "Judul Artikel"
                                    },
                                    "isi": {
                                        "description": "Isi artikel",
                                        "type": "string",
                                        "example": "Isi artikel..."
                                    },
                                    "foto": {
                                        "description": "Path atau URL ke file foto",
                                        "type": "string",
                                        "example": "path/to/foto.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Artikel created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Artikels"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Failed to create data"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error message details"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/artikels/{id}": {
            "get": {
                "tags": [
                    "Artikels"
                ],
                "summary": "Get artikel by ID",
                "description": "Returns a single artikel based on ID",
                "operationId": "getArtikelById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of artikel to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Artikels"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Artikel not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Data not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Failed to fetch data"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error message details"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Artikels"
                ],
                "summary": "Update an existing artikel",
                "description": "Updates an existing artikel and returns the updated artikel data",
                "operationId": "updateArtikel",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of artikel to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "judul": {
                                        "description": "Judul artikel, maksimum 255 karakter",
                                        "type": "string",
                                        "example": "Judul Artikel Baru"
                                    },
                                    "isi": {
                                        "description": "Isi artikel",
                                        "type": "string",
                                        "example": "Isi artikel yang diperbarui..."
                                    },
                                    "foto": {
                                        "description": "Path atau URL ke file foto",
                                        "type": "string",
                                        "example": "path/to/foto_baru.jpg",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Artikel updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Artikels"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Artikel not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Data not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Failed to update data"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error message details"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Artikels"
                ],
                "summary": "Delete an artikel",
                "description": "Deletes an artikel based on ID",
                "operationId": "deleteArtikel",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of artikel to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Artikel deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Artikel deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Artikel not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Data not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Failed to delete data"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error message details"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Artikels": {
                "title": "Artikel",
                "required": [
                    "judul",
                    "isi",
                    "foto"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "judul": {
                        "type": "string",
                        "example": "Judul Artikel"
                    },
                    "isi": {
                        "type": "string",
                        "example": "Isi lengkap artikel..."
                    },
                    "foto": {
                        "type": "string",
                        "format": "url",
                        "example": "http://example.com/image.jpg"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
...
}